Make Version::getRevision great again! (#530)
git recently introduced a change that short commit SHAs get longer to
avoid collisions. This bit Equalizer now, where git-describe delivers 9
byte of SHA.

+++ cpp/version.cpp
@@ -33,7 +33,7 @@ std::string Version::getString()
     return version.str();
 }
 
-int Version::getRevision()
+unsigned long long Version::getRevision()
 {
     return @NAMESPACE@_VERSION_REVISION;
 }
